+++ src/render/WavefrontIntersectionBackend.cpp
@@ -816,6 +816,9 @@ namespace render {
     if (hit.material < scene.materials().size()) {
       result.material = scene.materials()[hit.material];
     }
+    if (!result.material) {
+      result.material = primitive->material();
+    }
     result.hitPoint = HitPoint(
       primitive, hit.distance, Vector4d(hit.point[0], hit.point[1], hit.point[2], hit.point[3]),
       Vector3d(hit.normal[0], hit.normal[1], hit.normal[2]), Vector2d(hit.uv[0], hit.uv[1]));